+++ modules/app_eks/lb_controller/variables.tf
@@ -1,0 +1,10 @@
+variable "namespace" {
+   type = string
+}
+
+variable "oidc_provider" {
+  type = object({
+    arn = string
+    url = string
+  })
+}